+++ pulsar/consumer_partition.go
@@ -1670,7 +1670,7 @@ func (pc *partitionConsumer) reconnectToBroker(connectionClosed *connectionClose
 		}
 		pc.log.WithError(err).Error("Failed to create consumer at reconnect")
 		errMsg := err.Error()
-		if strings.Contains(errMsg, errTopicNotFount) {
+		if strings.Contains(errMsg, errMsgTopicNotFound) {
 			// when topic is deleted, we should give up reconnection.
 			pc.log.Warn("Topic Not Found.")
 			break