+++ pulsar/error_test.go
@@ -15,10 +15,22 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package utils
+package pulsar
 
-type TopicAutoCreationConfig struct {
-	Allow      bool      `json:"allowAutoTopicCreation"`
-	Type       TopicType `json:"topicType"`
-	Partitions *int      `json:"defaultNumPartitions"`
+import (
+	"errors"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_joinErrors(t *testing.T) {
+	err1 := errors.New("err1")
+	err2 := errors.New("err2")
+	err3 := errors.New("err3")
+	err := joinErrors(ErrInvalidMessage, err1, err2)
+	assert.True(t, errors.Is(err, ErrInvalidMessage))
+	assert.True(t, errors.Is(err, err1))
+	assert.True(t, errors.Is(err, err2))
+	assert.False(t, errors.Is(err, err3))
 }